Almost done with everything

+++ tools/equations.c
@@ -2,9 +2,9 @@
 #include <math.h>
 
 
-int calculate_distance(int x1, int y1, int x2, int y2) {
-    int part_1 = (int) (x1 - x2) * (x1 - x2);
-    int part_2 = (int) (y1 - y2) * (y1 - y2);
+double calculate_distance(int x1, int y1, int x2, int y2) {
+    double part_1 = (int) (x1 - x2) * (x1 - x2);
+    double part_2 = (int) (y1 - y2) * (y1 - y2);
 
-    return (int) sqrt(part_1 + part_2);
+    return (double) sqrt(part_1 + part_2);
 }
