+++ tools/include/equations.h
@@ -1,6 +1,6 @@
 #ifndef TERMINAL_SNAKE_GAME_EQUATIONS_H
 #define TERMINAL_SNAKE_GAME_EQUATIONS_H 1
 
-extern int calculate_distance(int x1, int y1, int x2, int y2);
+extern double calculate_distance(int x1, int y1, int x2, int y2);
 
 #endif
